pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pos-frontend"
version = "0.1.0"
description = "Front-end aspect of the POS pipeline for weekly transfers, reporting, and price correction"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pos-core-etl",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "statsmodels>=0.12.0",
    "openpyxl>=3.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=0.5.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
